Pass: Added codelogger logging of function definitions and domain inference steps using solver.  Fixed bug in loop partition.
Fixed bug that LoopPartition was not actually checking the partition information to be relevant to the current loop; in programs with multiple loops using the same
variable names, that could be a problem.  Mind you, variable names are made unique so not really an issue.
Extract solutions can now determine whether the solutions are inexact.